+++ multiboot2-header/src/builder/mod.rs
@@ -4,5 +4,5 @@ mod header;
 mod information_request;
 pub(self) mod traits;
 
-pub use header::Multiboot2HeaderBuilder;
+pub use header::HeaderBuilder;
 pub use information_request::InformationRequestHeaderTagBuilder;