+++ multiboot2-header/src/lib.rs
@@ -3,14 +3,14 @@
 //!
 //! # Example
 //! ```rust
-//! use multiboot2_header::builder::{InformationRequestHeaderTagBuilder, Multiboot2HeaderBuilder};
+//! use multiboot2_header::builder::{InformationRequestHeaderTagBuilder, HeaderBuilder};
 //! use multiboot2_header::{HeaderTagFlag, HeaderTagISA, MbiTagType, RelocatableHeaderTag, RelocatableHeaderTagPreference, Multiboot2Header};
 //!
 //! // Small example that creates a Multiboot2 header and parses it afterwards.
 //!
 //! // We create a Multiboot2 header during runtime here. A practical example is that your
 //! // program gets the header from a file and parses it afterwards.
-//! let mb2_hdr_bytes = Multiboot2HeaderBuilder::new(HeaderTagISA::I386)
+//! let mb2_hdr_bytes = HeaderBuilder::new(HeaderTagISA::I386)
 //!     .relocatable_tag(RelocatableHeaderTag::new(
 //!         HeaderTagFlag::Required,
 //!         0x1337,